Changed type of q_matrix and qp scale value

+++ src/oapv_def.h
@@ -125,7 +125,7 @@ struct oapv_fh {
     int       full_range_flag;                /* u( 1) */
     int       use_q_matrix;                   /* u( 1) */
     /* (start) quantization_matrix  */
-    int       q_matrix[N_C][OAPV_BLK_H][OAPV_BLK_W]; /* u( 8) */
+    u8        q_matrix[N_C][OAPV_BLK_H][OAPV_BLK_W]; /* u( 8) */
     /* ( end ) quantization_matrix  */
     /* (start) tile_info */
     int       tile_width_in_mbs;            /* u(20) */
